add tests for name
